Centering Auth Container
Some updates to packages for our project along with centering the authorization container. The way that the root page was originally set up, it was restricting half of the page from doing something, either displaying content or even registering as part of the webpage. Fixed by creating a style sheet for root page in app.tsx.